agents/adk-agent-pocs/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "adk_agent_samples"
version = "0.1.0"
description = "ADK Agent Samples"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "boto3>=1.42.59",
    "google-adk>=1.28.0",
    "google-genai>=1.69.0",
    "json-repair>=0.44.1",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-mcp-adapters>=0.1.0",
    "langgraph>=0.2.0",
    "langsmith",
]

[tool.hatch.build.targets.wheel]
packages = ["agents", "langgraph_agents", "shared"]

[tool.pytest.ini_options]
testpaths = ["agents", "langgraph_agents", "shared", "mcp_servers"]
pythonpath = ["."]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "debugpy>=1.8.20",
    "fastmcp>=2.13.0",
    "google-adk[eval]>=1.18.0",
    "pytest",
    "pytest-asyncio>=0.24.0",
    "python-dotenv>=1.2.1",
]
